+++ src/test/java/com/structurizr/export/plantuml/C4PlantUMLDiagramExporterTests.java
@@ -353,8 +353,8 @@ public class C4PlantUMLDiagramExporterTests extends AbstractExporterTests {
         relationship.addProperty("Prop1", "Value1");
         relationship.addProperty("Prop2", "Value2");
 
-        workspace.getViews().getConfiguration().addProperty(C4PlantUMLExporter.C4PLANTUML_ADD_ELEMENT_PROPERTIES_PROPERTY, Boolean.TRUE.toString());
-        workspace.getViews().getConfiguration().addProperty(C4PlantUMLExporter.C4PLANTUML_ADD_RELATIONSHIP_PROPERTIES_PROPERTY, Boolean.TRUE.toString());
+        workspace.getViews().getConfiguration().addProperty(C4PlantUMLExporter.C4PLANTUML_ELEMENT_PROPERTIES_PROPERTY, Boolean.TRUE.toString());
+        workspace.getViews().getConfiguration().addProperty(C4PlantUMLExporter.C4PLANTUML_RELATIONSHIP_PROPERTIES_PROPERTY, Boolean.TRUE.toString());
         ContainerView view = workspace.getViews().createContainerView(softwareSystem, "containerView", "");
         view.addDefaultElements();